trusted-firmware-m: trusted-firmware-m: bump to upstream TF-Mv1.2-RC2
Synchronize the trusted-firmware-m directory with the
corresponding upstream repository up to commit
368a25ff5e4925eed532fe7f6d03c3be9d8c4062
(TF-Mv1.2-RC2)

+++ trusted-firmware-m/platform/include/tfm_hal_ps.h
@@ -8,6 +8,8 @@
 #ifndef __TFM_HAL_PS_H__
 #define __TFM_HAL_PS_H__
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -25,7 +27,7 @@ extern "C" {
  * If an error is detected within this function, is should leave the
  * content of the parameters unchanged.
  */
-void tfm_hal_ps_fs_info(uint32_t *flash_area_addr, uint32_t *flash_area_size);
+void tfm_hal_ps_fs_info(uint32_t *flash_area_addr, size_t *flash_area_size);
 
 #ifdef __cplusplus
 }
